date time

+++ app/Http/Controllers/BookingController.php
@@ -199,6 +199,8 @@ class BookingController extends Controller
      */
     public function destroy(Reservation $booking, Request $request)
     {
+        abort_unless(auth()->user()?->getAllPermissions()->contains('name', 'delete booking'), 403);
+
         try {
             $booking->delete();
             $redirectParams = $this->getFilterParams($request);
